+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 更改日志 | Change Log
 
+## 2023-12-10
+
+- 使用 NoneBot 框架测试 QQ 协议适配器 | Test QQ protocol adapter with NoneBot framework
+
 ## 2023-12-09
 
 - 使用 qq-botpy 测试目前开放的权限 | Test the currently open permissions with qq-botpy